Refactoring and drop some component
