+++ src/org/tmpotter/core/BitextModel.java
@@ -26,7 +26,7 @@
  *
  * *************************************************************************/
 
-package org.tmpotter.engine;
+package org.tmpotter.core;
 
 import java.util.LinkedList;
